Implement zero-cost transposition
The tensor keeps a permutation order which is the only thing changed
when transposing. When the raw data is accessed, materialize_transpose()
can be used to actually transpose the raw data.

With this change, it makes more sense to transpose "in-place", i.e. not
to return a new tensor, because that would copy the data.
If the user wants the old tensor unchanged, it's better to leave it up
to them to copy the tensor and then transpose the copy, which makes the
data copy clear to them.