Add job dependency to prevent double docker build
Rebuild always builds when the Dockerfile is changed.
The cleanest solution would be a dependency of build on rebuild. Then,
on a change to the Dockerfile, rebuild would first build the new image
and build (which always runs) would see the image and do nothing.
However, as rebuild is not always running, build can not depend on it.
To prevent two docker builds at the same time, the dependency thus has
to be in opposite direction. Then, when e.g. a new branch with a
Dockerfile change is pushed, first the build runs, builds the new image,
writes a file as marker and then the rebuild runs, but doesn't build
again because it the marker file exists.
Without this file solution, two builds would be triggered.