Remove second Docker job again
After discussion, we agreed that the additional complexity of this job,
which is only required for creating images for branches that exist when
this MR is merged, is not worth it.
Instead, one can trigger an update manually by pushing some change to
the Dockerfile to have the pipeline build an image. Afterwards, the
commit can be dropped again.