Add an all_close method for tensors
Tensors do not directly implement PartialEq, because f64 don't do it (and, consequentially, complex numbers, too).
This makes sense, floats should be compared with a margin to allow for rounding errors.
Since there is no single margin that would work for all cases, we decided to not implement PartialEq, but allow an equality check with an explicitly provided margin.