Change shape to usize
The dimensions are bound by the length of the data vector, hence they are bound by usize::MAX.
Imagine this library being used on a 16-bit architecture and a 128-bit architecture. Would it make sense that the shape is always u64, even though less (or more) elements can be allocated in the data vector? In my opinion, no.