Consume tensors in contract
This enables us to early-drop them in the contraction. Without it, we need memory for A, B, A^T, B^T and C. However, A can directly be dropped after we computed A^T and B can be dropped after we computed B^T, thus decreasing the total memory needed.