Prevent clone for shared data
The previous approach was already pretty good: The tensors were consumed and, if contiguous, the data was returned or else the transpose was computed and the original data dropped.

However, `unwrap_or_clone` cloned the data if it was shared. We only need read-only access, so we don't care if the data is shared or not, meaning we don't need the clone. We can not return a reference or a Cow, because the tensor dies at the end of the method, so we have to make sure the data - shared or not - lives on until we are done with the contraction. Moving the Arc out allows for this temporary ownership.

Here are the four scenarios:
- not contigous, not shared: Transpose is computed, original data dropped at end of method
- not contiguous, shared: Tranpose is computed, handle to original data dropped at end of method
- contigous, not shared: Handle to data returned, data dropped at end of contract
- contigous, shared: Handle of data returned, handle dropped at end of contract